feat: abstract lean client setup for websocket clients
